ZON-6478: Move content query functionality to own module zeit.contentquery
