+++ core/src/zeit/content/cp/ftesting.zcml
@@ -28,6 +28,8 @@
 
   <include package="zeit.content.gallery" />
 
+  <include package="zeit.contentquery" />
+
   <!-- Allow test user to create and move areas -->
   <grant
     principal="zope.Authenticated"
