+++ core/src/zeit/contentquery/__init__.py
@@ -1,0 +1,1 @@
+# python package
